inner-884-bugfix: prepare phase cause error

+++ src/main/java/com/actiontech/dble/backend/mysql/ByteUtil.java
@@ -1,8 +1,8 @@
 /*
-* Copyright (C) 2016-2021 ActionTech.
-* based on code by MyCATCopyrightHolder Copyright (c) 2013, OpenCloudDB/MyCAT.
-* License: http://www.gnu.org/licenses/gpl.html GPL version 2 or higher.
-*/
+ * Copyright (C) 2016-2021 ActionTech.
+ * based on code by MyCATCopyrightHolder Copyright (c) 2013, OpenCloudDB/MyCAT.
+ * License: http://www.gnu.org/licenses/gpl.html GPL version 2 or higher.
+ */
 package com.actiontech.dble.backend.mysql;
 
 /**
@@ -29,7 +29,7 @@ public final class ByteUtil {
         long l = data[offset] & 0xff;
         l |= (data[++offset] & 0xff) << 8;
         l |= (data[++offset] & 0xff) << 16;
-        l |= (data[++offset] & 0xff) << 24;
+        l |= ((long) (data[++offset] & 0xff)) << 24;
         return l;
     }
